untaint upper 32 bits of 32-bit destination registers
This patch untaints the upper 32 bits of 32-bit destination registers.
On x86, writes to 32-bit subregisters (e.g., EAX) implicitly zero the
upper 32 bits of the full 64-bit register (e.g., RAX). Thus, it is
safe for SPT to unconditionally untaint the upper 32 bits of the full
register whenever the lower 32 bits are written to (e.g., by the
instruction `mov eax,42`).